% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{category_search}
\alias{category_search}
\title{Allows a user to see their category options.
Defaults to all options, or shows those that contain the provided query as a substring.}
\usage{
category_search(query = "")
}
\arguments{
\item{query}{A substring of the category name to match.}
}
\value{
Category titles and Ids matching the query, as a data frame.
}
\description{
Allows a user to see their category options.
Defaults to all options, or shows those that contain the provided query as a substring.
}
\examples{
category_search("pants")

}
