% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_search.R
\name{clean_data}
\alias{clean_data}
\title{Cleans Jeopardy datatable for display.}
\usage{
clean_data(data)
}
\arguments{
\item{data}{The dataset to manage}
}
\value{
A clean data frame for the question_search function
}
\description{
Cleans Jeopardy datatable for display.
}
