% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{display_question}
\alias{display_question}
\title{Displays the question in a standardized format}
\usage{
display_question(question)
}
\arguments{
\item{question}{a question as a row of a data frame.}
}
\value{
None
}
\description{
Displays the question in a standardized format
}
