% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_jeopardy.R
\name{get_board}
\alias{get_board}
\title{Builds a question board with random categories}
\usage{
get_board(n_categories)
}
\arguments{
\item{n_categories}{number of categories on the board}
}
\value{
a questions board as a data frame
}
\description{
Builds a question board with random categories
}
