% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{get_q}
\alias{get_q}
\title{Gets question based off value and category from API}
\usage{
get_q(value, category)
}
\arguments{
\item{value}{pts the question is worth (int)}

\item{category}{category the question will belong to (int)}
}
\description{
Gets question based off value and category from API
}
