% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{get_trivia_qs}
\alias{get_trivia_qs}
\title{Gets random questions from the API}
\usage{
get_trivia_qs(num_questions = 1)
}
\arguments{
\item{num_questions}{number of question you want returned}
}
\value{
returns a vector of q's
}
\description{
Gets random questions from the API
}
