% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_search.R
\name{jeopardy_question_search}
\alias{jeopardy_question_search}
\title{Creates an interactive table of Jeopardy questions
following given criteria, limited at 100 questions}
\usage{
jeopardy_question_search(
  value = NULL,
  category = NULL,
  min_date = NULL,
  max_date = NULL
)
}
\arguments{
\item{value}{The in-game dollar value that corresponds to the Jeopardy quesion}

\item{category}{The category of the question, as an id integer}

\item{min_date}{The earliest date to show, based on original air date}

\item{max_date}{The latest date to show, based on original air date}
}
\value{
An data frame of the specified questions, displays as data table in viewer.
}
\description{
Creates an interactive table of Jeopardy questions
following given criteria, limited at 100 questions
}
