% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeopardy_board.R
\name{play_jeopardy}
\alias{play_jeopardy}
\title{Displays a jeopardy board that a user can select questions from}
\usage{
play_jeopardy(n_categories = 6)
}
\arguments{
\item{n_categories}{How many categories in the Jeopardy table. Default is 6.}
}
\value{
None
}
\description{
Displays a jeopardy board that a user can select questions from
}
