% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jeopardy_board.R
\name{play_one_question}
\alias{play_one_question}
\title{Displays a jeopardy board that a user can select questions from}
\usage{
play_one_question(questions, category_ids)
}
\arguments{
\item{questions}{questions board.}

\item{category_ids}{id numbers of categories corresponding to questions board columns.}
}
\value{
updated questions board
}
\description{
Displays a jeopardy board that a user can select questions from
}
