% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{query_categories}
\alias{query_categories}
\title{Queries the API for categories given an offset}
\usage{
query_categories(offset)
}
\arguments{
\item{offset}{offsets the returned categories. Useful in pagination}
}
\value{
100 categories as a data frame
}
\description{
Queries the API for categories given an offset
}
