% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random_jeopardy}
\alias{random_jeopardy}
\title{Gets random jeopardy questions from the API}
\usage{
random_jeopardy(num_questions = 1)
}
\arguments{
\item{num_questions}{number of question you want returned}
}
\value{
returns a vector of q's
}
\description{
Gets random jeopardy questions from the API
}
