% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{wait}
\alias{wait}
\title{Waits for enter key before continuing}
\usage{
wait(to = "")
}
\arguments{
\item{to}{what will happen after you press enter. Default empty string.}
}
\value{
None
}
\description{
Waits for enter key before continuing
}
