% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_jeopardy.R
\name{play_jeopardy}
\alias{play_jeopardy}
\title{Interactive game of jeopardy, where user can select which value/category
questions they'd like to answer.}
\usage{
play_jeopardy(n_categories = 6)
}
\arguments{
\item{n_categories}{How many categories in the Jeopardy table. Default is 6.}
}
\value{
None
}
\description{
Interactive game of jeopardy, where user can select which value/category
questions they'd like to answer.
}
