% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play_jeopardy.R
\name{play_one_question}
\alias{play_one_question}
\title{Goes through one round of jeopardy
selecting category and value, showing question, and showing answer}
\usage{
play_one_question(questions, category_ids)
}
\arguments{
\item{questions}{questions board as a data frame.}

\item{category_ids}{id numbers of categories corresponding to questions board columns.}
}
\value{
updated questions board as a data frame
}
\description{
Goes through one round of jeopardy
selecting category and value, showing question, and showing answer
}
